refactor(pos): replace Web Serial with CSS print for USB thermal printers
3nStar RPT printers connect as USB Printer Class (not CDC/serial),
so Web Serial API cannot reach them. The correct approach is window.print()
with proper 80mm @page CSS, which routes through the OS printer driver.

- Remove useThermalPrinter hook, thermal-printer.ts, thermal-printer.types.d.ts
- Uninstall @point-of-sale/receipt-printer-encoder
- Extract PosReceiptContent component (shared between dialog and hidden DOM node)
- Render receipt in off-screen hidden div so window.print() works without dialog
- Auto-print via requestAnimationFrame after checkout when autoPrintReceipt is on
- Improve @media print CSS: @page size 80mm auto, margin 0, monospace font 9pt
- Remove Web Serial i18n keys, simplify PosReceiptPreview
