+++ app/routes/pos/terminal.tsx
@@ -24,8 +24,7 @@ import { PosCloseShiftDialog } from '~/features/pos/components/PosCloseShiftDial
 import { PosCashMovementDialog } from '~/features/pos/components/PosCashMovementDialog'
 import { PosCreateCustomerDialog } from '~/features/pos/components/PosCreateCustomerDialog'
 import type { ReceiptData } from '~/features/pos/components/PosReceiptPreview'
-import { useThermalPrinter } from '~/features/pos/hooks/useThermalPrinter'
-import { buildReceiptBytes } from '~/features/pos/lib/thermal-printer'
+import { PosReceiptContent } from '~/features/pos/components/PosReceiptContent'
 import {
   type CartItem,
   type PosProductForGrid,
@@ -344,7 +343,6 @@ export default function PosTerminal({ loaderData }: Route.ComponentProps) {
   } = loaderData
 
   const { t } = useTranslation()
-  const thermalPrinter = useThermalPrinter()
 
   useEffect(() => {
     if (autoClosedStaleSession) {
@@ -617,18 +615,14 @@ export default function PosTerminal({ loaderData }: Route.ComponentProps) {
     setSelectedCustomer(null)
     setPaymentOpen(false)
 
-    const printerReady =
-      thermalPrinter.status === 'connected' || thermalPrinter.status === 'printing'
-
-    if (terminal.autoPrintReceipt && printerReady) {
-      // Auto-print: send bytes directly, skip dialog
-      thermalPrinter
-        .print(buildReceiptBytes(receipt))
-        .then(() => toast.success(t('pos.printSuccess')))
-        .catch(() => {
-          // If auto-print fails, fall back to showing the receipt dialog
-          setReceiptOpen(true)
+    if (terminal.autoPrintReceipt) {
+      // Auto-print: the receipt is already rendered in the hidden DOM node.
+      // Two rAF frames ensure React has committed the new receiptData before printing.
+      requestAnimationFrame(() => {
+        requestAnimationFrame(() => {
+          window.print()
         })
+      })
     } else {
       setReceiptOpen(true)
     }
@@ -735,7 +729,6 @@ export default function PosTerminal({ loaderData }: Route.ComponentProps) {
         open={receiptOpen}
         onOpenChange={setReceiptOpen}
         receipt={receiptData}
-        printer={thermalPrinter}
       />
 
       <PosOpenShiftDialog
@@ -767,16 +760,39 @@ export default function PosTerminal({ loaderData }: Route.ComponentProps) {
         isSubmitting={fetcher.state === 'submitting'}
       />
 
+      {/*
+        Hidden receipt node — always present in the DOM when there is receiptData.
+        window.print() targets #pos-receipt via the @media print CSS below,
+        so auto-print works without opening the dialog.
+      */}
+      {receiptData && (
+        <div
+          aria-hidden
+          style={{ position: 'absolute', left: '-9999px', top: 0, pointerEvents: 'none' }}
+        >
+          <PosReceiptContent receipt={receiptData} />
+        </div>
+      )}
+
       <style>{`
         @media print {
-          body * { visibility: hidden; }
-          #pos-receipt, #pos-receipt * { visibility: visible; }
+          @page {
+            size: 80mm auto;
+            margin: 0;
+          }
+          body * { visibility: hidden !important; }
+          #pos-receipt, #pos-receipt * { visibility: visible !important; }
           #pos-receipt {
-            position: absolute;
-            left: 0;
+            position: fixed;
             top: 0;
-            width: 80mm;
-            padding: 4mm;
+            left: 0;
+            width: 76mm;
+            padding: 2mm;
+            font-family: 'Courier New', Courier, monospace;
+            font-size: 9pt;
+            line-height: 1.4;
+            color: #000;
+            background: #fff;
           }
         }
       `}</style>
