+++ drizzle/meta/_journal.json
@@ -148,6 +148,13 @@
       "when": 1772585193822,
       "tag": "0020_sparkling_thunderbolt",
       "breakpoints": true
+    },
+    {
+      "idx": 21,
+      "version": "7",
+      "when": 1772730816746,
+      "tag": "0021_tough_revanche",
+      "breakpoints": true
     }
   ]
 }
