+++ package.json
@@ -36,7 +36,6 @@
   },
   "dependencies": {
     "@neondatabase/serverless": "^1.0.2",
-    "@point-of-sale/receipt-printer-encoder": "^3.0.3",
     "@radix-ui/react-alert-dialog": "^1.1.15",
     "@radix-ui/react-avatar": "^1.1.11",
     "@radix-ui/react-checkbox": "^1.3.3",
